Apply SVUnit plugin
This is a bit more involved, because we have to add a dependency on the SVUnit library, which doesn't yet provide a Gradle build. We can only inject plugins from included builds, so we have to wrap the `gradle-hdvl.svunit-build` plugin.

Using a project as a source dependency will fail when that project has included builds, so we have to take care to only include the plugin build when we're building the project standalone.