Pass output file descriptor when constructing a `dict_writer`
We pass a file descriptor instead of just a file path. While we could, when given a file path, open the file in the constructor, there's no clear point where we would close the file. We have to leave closing to the surrounding code, but this means we have to relinquish opening as well.