Drop redundant test
We needed this test when we wanted to implement the quoting of values in subsequent (i.e. not the first) row locations. Right now, this test had to also be updated when we changed the behavior to not quote anymore.

Removing this test does sort of leave a gap because we're not testing that the first element is handled the same as subsequent elements. At the same time, if we were to want to change the code to treat the two cases differently, we would anyway have to write a new test. We're probably fine after dropping it.